Safebooru: update restricted tags
* remove "censored": this should be already handled by rating, as
censored genitals, sexual acts etc should always be q when they're
explicit enough that they're being censored. There's also a lot of false
positives in keeping this in, such as identity censor, censored
foods/innocuous items for comedy's sake, etc.

* add "sex_toy" - this is a recently populated tag since all types of
sex toys now implicate it, and adding this also makes sure we cover
the "sexual items" part of censored content that might be rating:safe
but not something we want on safebooru, and would filter through if
"censored" is not completely restricted